+++ src/main/java/com/example/demo/services/impl/CRUDEmployeeService.java
@@ -3,8 +3,9 @@ package com.example.demo.services.impl;
 import com.example.demo.dtos.EmployeeDTO;
 import com.example.demo.entities.Company;
 import com.example.demo.entities.Employee;
+import com.example.demo.exceptions.NoContentPresentException;
+import com.example.demo.exceptions.ResourceNotFoundException;
 import com.example.demo.mappers.EmployeeMapper;
-import com.example.demo.repositories.CompanyRepository;
 import com.example.demo.repositories.EmployeeRepository;
 import com.example.demo.services.CRUDService;
 import lombok.AccessLevel;
@@ -26,8 +27,6 @@ public class CRUDEmployeeService implements CRUDService<EmployeeDTO> {
 
     EmployeeRepository employeeRepository;
 
-    CompanyRepository companyRepository;
-
     @Override
     public List<EmployeeDTO> getAll() {
         log.info("Getting all employees");
@@ -41,6 +40,9 @@ public class CRUDEmployeeService implements CRUDService<EmployeeDTO> {
     public EmployeeDTO getById(Long id) {
         log.info("Getting employee by by id: {}", id);
         Optional<Employee> employee = employeeRepository.findById(id);
+        if (employee.isEmpty()) {
+            throw new ResourceNotFoundException("Cannot find employee with id: " + id);
+        }
         return employee.map(employeeMapper::toDto).get();
     }
 
@@ -48,31 +50,32 @@ public class CRUDEmployeeService implements CRUDService<EmployeeDTO> {
     public EmployeeDTO create(EmployeeDTO employeeDTO) {
         log.info("Creating new employee.");
         Employee employee = employeeMapper.toEntity(employeeDTO);
-        Optional<Company> company = companyRepository.findByName(employeeDTO.getCompany().getName());
-        employee.setCompany(company.get());
-        employeeRepository.save(employee);
-        return employeeDTO;
+        Employee savedEmployee = employeeRepository.insertEmployee(employee.getName(), employee.getSurname(),
+                employee.getSalary(), employee.getHiringDate(), employee.getJob().toString(),
+                employee.getCompany().getName());
+        return employeeMapper.toDto(savedEmployee);
     }
 
     @Transactional
     @Override
     public void update(EmployeeDTO employeeDTO, Long id) {
         log.info("Updating employee with id: {}", id);
-        Employee employee = employeeMapper.toEntity(employeeDTO);
-        Optional<Long> companyId = Optional.empty();
-        if(employeeDTO.getCompany()!=null){
-            Optional<Company> company = companyRepository.findByName(employeeDTO.getCompany().getName());
-            if (company.isPresent()) {
-                companyId = Optional.of(company.get().getId());
-            }
+        if (!employeeRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Cannot find employee with id: " + id);
         }
+        Employee employee = employeeMapper.toEntity(employeeDTO);
+        Optional<Company> company = Optional.of(employee.getCompany());
         employeeRepository.updateEmployee(id, employee.getName(), employee.getSurname(),
-                employee.getSalary(), employee.getHiringDate(), employeeDTO.getJob(), companyId.orElse(null));
+                employee.getSalary(), employee.getHiringDate(), employeeDTO.getJob(),
+                company.map(Company::getName).orElse(null));
     }
 
     @Override
     public void delete(Long id) {
         log.info("Deleting employee with id: {}", id);
+        if (!employeeRepository.existsById(id)) {
+            throw new NoContentPresentException("No content present with id: " + id);
+        }
         employeeRepository.deleteById(id);
     }
 }
